move models.py to main_app
